Rework adapter structure
