--- src/adapter/10/renderer/inspectVNode.ts
@@ -1,80 +1,0 @@
-import {
-	getComponent,
-	getComponentHooks,
-	getDisplayName,
-	getSuspenseStateKey,
-	getVNodeParent,
-	isSuspenseVNode,
-} from "../vnode";
-import { serialize, cleanContext, cleanProps } from "../utils";
-import { RendererConfig10, getDevtoolsType } from "../renderer";
-import { ID } from "../../../view/store/types";
-import { IdMappingState, getVNodeById } from "../IdMapper";
-import { Options, VNode } from "preact";
-import { inspectHooks } from "./hooks";
-import { InspectData } from "../../adapter/adapter";
-
-/**
- * Serialize all properties/attributes of a `VNode` like `props`, `context`,
- * `hooks`, and other data. This data will be requested when the user selects a
- * `VNode` in the devtools. It returning information will be displayed in the
- * sidebar.
- */
-export function inspectVNode(
-	ids: IdMappingState,
-	config: RendererConfig10,
-	options: Options,
-	id: ID,
-	supportsHooks: boolean,
-): InspectData | null {
-	const vnode = getVNodeById(ids, id);
-	if (!vnode) return null;
-
-	const c = getComponent(vnode);
-	const hasState =
-		typeof vnode.type === "function" &&
-		c != null &&
-		Object.keys(c.state).length > 0;
-
-	const hasHooks = c != null && getComponentHooks(c) != null;
-	const hooks =
-		supportsHooks && hasHooks ? inspectHooks(config, options, vnode) : null;
-	const context = c != null ? serialize(config, cleanContext(c.context)) : null;
-	const props =
-		vnode.type !== null ? serialize(config, cleanProps(vnode.props)) : null;
-	const state = hasState ? serialize(config, c!.state) : null;
-
-	let suspended = false;
-	let canSuspend = false;
-	let item: VNode | null = vnode;
-	while (item) {
-		if (isSuspenseVNode(item)) {
-			canSuspend = true;
-
-			const c = getComponent(item);
-			if (c) {
-				const key = getSuspenseStateKey(c);
-				if (key) {
-					suspended = !!(c as any)._nextState[key];
-				}
-			}
-			break;
-		}
-
-		item = getVNodeParent(item);
-	}
-
-	return {
-		context,
-		canSuspend,
-		key: vnode.key || null,
-		hooks: supportsHooks ? hooks : !supportsHooks && hasHooks ? [] : null,
-		id,
-		name: getDisplayName(vnode, config),
-		props,
-		state,
-		// TODO: We're not using this information anywhere yet
-		type: getDevtoolsType(vnode),
-		suspended,
-	};
-}
