+++ src/adapter/shared/serialize.test.tsx
@@ -1,9 +1,41 @@
+import { expect } from "chai";
+import { cleanContext, cleanProps, jsonify } from "./serialize";
 import { h, Component, createContext, render } from "preact";
 import { teardown } from "preact/test-utils";
 import { setupScratch } from "./renderer.test";
-import { expect } from "chai";
-import { cleanContext, cleanProps } from "./utils";
-import { getActualChildren } from "./vnode";
+import { getActualChildren } from "../10/bindings";
+
+describe("jsonify", () => {
+	it("should clean circular references", () => {
+		const data1: any = { foo: 123 };
+		data1.foo = data1;
+		expect(jsonify(data1, () => null, new Set())).to.deep.equal({
+			foo: "[[Circular]]",
+		});
+
+		const data2: any = { foo: [] };
+		data2.foo.push(data2);
+		expect(jsonify(data2, () => null, new Set())).to.deep.equal({
+			foo: ["[[Circular]]"],
+		});
+	});
+
+	it("should not treat values as circular", () => {
+		const data: any = { foo: 123, bar: { foo: 123 } };
+		expect(jsonify(data, () => null, new Set())).to.deep.equal({
+			foo: 123,
+			bar: { foo: 123 },
+		});
+	});
+
+	it("should parse symbols", () => {
+		const data: any = Symbol("foo");
+		expect(jsonify(data, () => null, new Set())).to.deep.equal({
+			type: "symbol",
+			name: "Symbol(foo)",
+		});
+	});
+});
 
 describe("cleanProps", () => {
 	let scratch: HTMLDivElement;
